type of GfxElement->colortable was changed to UINT32 in MAME37b12, thus had to change cursorcolortable[] to UINT32 too

+++ mess/vidhrdw/ted7360.c
@@ -498,7 +498,7 @@ static int rasterline = 0, lastline = 0;
 static double rastertime;
 static void ted7360_drawlines (int first, int last);
 
-static unsigned short cursorcolortable[2] =
+static UINT32 cursorcolortable[2] =
 {0};
 static struct GfxLayout cursorlayout =
 {
